simplify datamodel organization. Create "core" objects
