--- biosimulations/libs/datamodel/core/src/lib/aliases/Users.ts
@@ -1,2 +1,0 @@
-export type UserName = string;
-export type User = UserName;
