--- biosimulations/libs/datamodel/core/src/lib/aliases/formats.ts
@@ -1,1 +1,0 @@
-export type DateString = string;
