--- biosimulations/libs/datamodel/core/src/lib/aliases/identity.ts
@@ -1,4 +1,0 @@
-export type BiosimulationsId = string;
-export type KisaoId = string;
-export type EdamId = string;
-export type DOI = string;
