--- biosimulations/libs/datamodel/core/src/lib/common/biomodel-variable.dto.ts
@@ -1,13 +1,0 @@
-import { DTO } from '@biosimulations/datamodel/utils';
-import { PrimitiveType } from '../..';
-
-export class ModelVariableCore {
-  target: string;
-  group: string;
-  id: string;
-  name: string;
-  description: string;
-  type: PrimitiveType;
-  units: string;
-}
-export type BiomodelVariableDTO = DTO<ModelVariableCore>;
