--- biosimulations/libs/datamodel/core/src/lib/common/chart-data-field.dto.ts
@@ -1,11 +1,0 @@
-import { ChartDataFieldShape } from '../enums/chart-data-field-shape';
-import { ChartDataFieldType } from '../enums/chart-data-field-type';
-import { DTO } from '@biosimulations/datamodel/utils';
-
-export interface ChartDataFieldCore {
-  name?: string;
-  shape?: ChartDataFieldShape;
-  type?: ChartDataFieldType;
-}
-
-export type ChartDataFieldDTO = DTO<ChartDataFieldCore>;
