--- biosimulations/libs/datamodel/core/src/lib/common/identifier.dto.ts
@@ -1,8 +1,0 @@
-import { DTO } from '@biosimulations/datamodel/utils';
-
-// Identifiers.org identifier
-export interface IdentifierCore {
-  namespace: string;
-  identifier: string;
-}
-export type IdentifierDTO = DTO<IdentifierCore>;
