+++ biosimulations/libs/datamodel/core/src/lib/common/index.ts
@@ -1,20 +1,9 @@
-export { PersonDTO } from './person.dto';
-export { isAlgorithmParameterDTO } from './algorithm-parameter.dto';
-export { isBiomodelParameterDTO } from './biomodel-parameter.dto';
-export { IdentifierDTO } from './identifier.dto';
-export { AlgorithmDTO } from './algorithm.dto';
-export { AlgorithmParameterDTO } from './algorithm-parameter.dto';
-export { FormatDTO } from './format.dto';
-export { JournalReferenceDTO } from './journalreference.dto';
-export { BiomodelParameterDTO } from './biomodel-parameter.dto';
-export { BiomodelVariableDTO } from './biomodel-variable.dto';
-export { OntologyTermDTO } from './ontology-term.dto';
-export { ParameterChangeDTO } from './parameter-change.dto';
-export { ProjectProductDTO } from './project-product.dto';
-export { ChartDataFieldDTO } from './chart-data-field.dto';
-export { LicenseInfo } from './license-info';
-export { SimulationResultDTO } from './simulation-result-dto';
-export { TimePointDTO } from './time-point.dto';
-export { UserDTO } from './user.dto';
-export { VisualizationDataFieldDTO } from './visualization-data-field.dto';
-export { VisualizationLayoutElementDTO } from './visualization-layout-element.dto';
+export * from './alias';
+export * from './format';
+export * from './journalreference';
+export * from './license';
+export * from './format';
+export * from './ontology';
+export * from './person';
+export * from './taxon';
+export * from './primitive-type';
