+++ biosimulations/libs/datamodel/core/src/lib/common/journalreference.ts
