--- biosimulations/libs/datamodel/core/src/lib/common/license-info.ts
@@ -1,9 +1,0 @@
-import { License } from '../enums/licence';
-
-export interface LicenseInfo {
-  value: License;
-  name: string;
-  version: string;
-  swoId: number;
-  url: string;
-}
