--- biosimulations/libs/datamodel/core/src/lib/common/log-item.dto.ts
@@ -1,7 +1,0 @@
-import { DateString } from '../aliases/formats';
-
-export class LogItemDTO {
-  time: DateString;
-  type: string;
-  message: string;
-}
