--- biosimulations/libs/datamodel/core/src/lib/common/parameter-change.dto.ts
@@ -1,7 +1,0 @@
-import { AlgorithmParameterDTO } from './algorithm-parameter.dto';
-import { BiomodelParameterDTO } from './biomodel-parameter.dto';
-
-export interface ParameterChangeDTO {
-  parameter: BiomodelParameterDTO | AlgorithmParameterDTO;
-  value: number;
-}
