+++ biosimulations/libs/datamodel/core/src/lib/common/primitive-type.ts
@@ -1,0 +1,1 @@
+export type PrimitiveType = 'boolean' | 'integer' | 'float' | 'string';
