--- biosimulations/libs/datamodel/core/src/lib/common/time-point.dto.ts
@@ -1,8 +1,0 @@
-import { DTO } from '@biosimulations/datamodel/utils';
-
-export interface TimePointCore {
-  time: number;
-  value: number;
-}
-
-export type TimePointDTO = DTO<TimePointCore>;
