--- biosimulations/libs/datamodel/core/src/lib/common/visualization-data-field.dto.ts
@@ -1,7 +1,0 @@
-import { ChartDataFieldDTO } from './chart-data-field.dto';
-import { SimulationResultDTO } from './simulation-result-dto';
-
-export class VisualizationDataFieldDTO {
-  dataField: ChartDataFieldDTO;
-  simulationResults: SimulationResultDTO[];
-}
