--- biosimulations/libs/datamodel/core/src/lib/common/visualization-layout-element.dto.ts
@@ -1,7 +1,0 @@
-import { ChartDTO } from '../resources/chart.dto';
-import { VisualizationDataFieldDTO } from './visualization-data-field.dto';
-
-export class VisualizationLayoutElementDTO {
-  chart: ChartDTO;
-  data: VisualizationDataFieldDTO[];
-}
