+++ biosimulations/libs/datamodel/core/src/lib/core/biomodel.ts
@@ -1,6 +1,11 @@
+import {
+  IdentifierDTO,
+  PrimitiveType,
+  OntologyTermDTO,
+  FormatDTO,
+} from '../..';
 import { DTO } from '@biosimulations/datamodel/utils';
-import { IdentifierDTO } from './identifier.dto';
-import { PrimitiveType } from '../enums';
+import { TaxonDTO } from '../common/taxon';
 
 export interface BiomodelParameterCore {
   target: string;
@@ -21,3 +26,23 @@ export const isBiomodelParameterDTO = (
   param: any,
 ): param is BiomodelParameterDTO =>
   'units' in param && 'id' in param && 'name' in param && 'value' in param;
+
+export class BiomodelVariableCore {
+  target: string;
+  group: string;
+  id: string;
+  name: string;
+  description: string;
+  type: PrimitiveType;
+  units: string;
+}
+export type BiomodelVariableDTO = DTO<BiomodelVariableCore>;
+
+export class Biomodel {
+  taxon: TaxonDTO;
+  parameters: BiomodelParameterDTO[];
+  variables: BiomodelVariableDTO[];
+  file: string;
+  framework: OntologyTermDTO;
+  format: FormatDTO;
+}
