+++ biosimulations/libs/datamodel/core/src/lib/core/project.ts
@@ -1,3 +1,7 @@
+import { DTO } from '@biosimulations/datamodel/utils';
+import { JournalReferenceDTO } from '../common';
+import { BiosimulationsId } from '../common/alias';
+
 export enum ProjectProductType {
   figure = 'figure',
   table = 'table',
@@ -15,3 +19,12 @@ export const projectProductTypes = [
   { value: ProjectProductType.algorithm, name: 'algorithm' },
   { value: ProjectProductType.other, name: 'other' },
 ];
+
+export class ProjectProductCore {
+  reference: JournalReferenceDTO;
+  type: ProjectProductType;
+  label: string;
+  description: string;
+  resources: BiosimulationsId[];
+}
+export type ProjectProductDTO = DTO<ProjectProductCore>;
