--- biosimulations/libs/datamodel/core/src/lib/enums/access-level.ts
@@ -1,10 +1,0 @@
-export enum AccessLevel {
-  private = 'private',
-  public = 'public',
-  shared = 'shared',
-}
-export const accessLevels = [
-  { value: AccessLevel.private, name: 'private' },
-  { value: AccessLevel.public, name: 'public' },
-  { value: AccessLevel.shared, name: 'shared' },
-];
