--- biosimulations/libs/datamodel/core/src/lib/enums/chart-data-field-shape.ts
@@ -1,4 +1,0 @@
-export enum ChartDataFieldShape {
-  scalar = 'scalar',
-  array = 'array',
-}
