--- biosimulations/libs/datamodel/core/src/lib/enums/chart-data-field-type.ts
@@ -1,4 +1,0 @@
-export enum ChartDataFieldType {
-  dynamicSimulationResult = 'dynamicSimulationResult',
-  static = 'static',
-}
