--- biosimulations/libs/datamodel/core/src/lib/enums/chart-type-type.ts
@@ -1,4 +1,0 @@
-export enum ChartTypeType {
-  basic = 'basic',
-  advanced = 'advanced',
-}
