--- biosimulations/libs/datamodel/core/src/lib/enums/index.ts
@@ -1,12 +1,0 @@
-export { SimulationResultsFormat } from './simulation-results-format';
-export {
-  projectProductTypes,
-  ProjectProductType,
-} from './project-product-type';
-
-export { AccessLevel, accessLevels } from './access-level';
-export { ChartDataFieldShape } from './chart-data-field-shape';
-export { ChartDataFieldType } from './chart-data-field-type';
-export { ChartTypeType } from './chart-type-type';
-export { License } from './licence';
-export { ResourceType } from './resource-type';
