--- biosimulations/libs/datamodel/core/src/lib/enums/licence.ts
@@ -1,9 +1,0 @@
-export enum License {
-  cc0 = 'CC0',
-  cc_by = 'CC BY',
-  cc_by_sa = 'CC BY-SA',
-  cc_by_nc = 'CC BY-NC',
-  cc_by_nc_sa = 'CC BY-NC-SA',
-  mit = 'MIT',
-  other = 'Other',
-}
