--- biosimulations/libs/datamodel/core/src/lib/enums/resource-type.ts
@@ -1,9 +1,0 @@
-export enum ResourceType {
-  project = 'project',
-  model = 'model',
-  simulation = 'simulations',
-  chart = 'chart',
-  visualization = 'visualization',
-  file = 'file',
-  user = 'user',
-}
