--- biosimulations/libs/datamodel/core/src/lib/enums/simulation-results-format.ts
@@ -1,8 +1,0 @@
-export enum SimulationResultsFormat {
-  csv = 'csv',
-  hdf5 = 'hdf5',
-  n5 = 'n5',
-  numl = 'numl',
-  tsv = 'tsv',
-  zarr = 'zarr',
-}
